Do not keep a reference to the controller
We contextualise template rendering with the current request by storing
the controller in the template provider. This means that we keep hold of
the full request context until that reference is overwritten, either by
a subsequent hit using the same renderer or by the server shutting down.

This can result in unexpected cleanup behaviour. Destructors for objects
in the stash would normally be executed at the end of the request, but
when rendered with TT they get invoked in the middle of the next
TT-rendered request.

Weaken the ctx attribute value in addition to the renderer, so that it
will go out of scope when the main request cycle is done.